Backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
greenlet==3.0.3
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
alembic==1.13.1
python-dotenv==1.0.1
faker==24.4.0